+++ test/Masa.Framework.IntegrationTests.EventBus/Application/Command/RegisterUserCommand.cs
@@ -1,9 +1,9 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.EventBus.IntegrationTests.Application.Events;
+namespace Masa.Framework.IntegrationTests.EventBus.Application.Command;
 
-public record RegisterUserEvent : Event
+public record RegisterUserCommand : BuildingBlocks.ReadWriteSpliting.Cqrs.Commands.Command
 {
     public string Name { get; set; }
