+++ test/Masa.Framework.IntegrationTests.EventBus/Application/Events/AddGoodsIntegrationEvent.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.EventBus.IntegrationTests.Application.Events;
+namespace Masa.Framework.IntegrationTests.EventBus.Application.Events;
 
 public record AddGoodsIntegrationEvent : IntegrationEvent
 {
