+++ test/Masa.Framework.IntegrationTests.EventBus/Application/Events/RegisterUserEvent.cs
@@ -1,16 +1,11 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.EventBus.IntegrationTests.Domain.Aggregate;
+namespace Masa.Framework.IntegrationTests.EventBus.Application.Events;
 
-public class User : AggregateRoot<Guid>
+public record RegisterUserEvent : Event
 {
     public string Name { get; set; }
 
     public int Age { get; set; }
-
-    public User()
-    {
-        Id = Guid.NewGuid();
-    }
 }
