+++ test/Masa.Framework.IntegrationTests.EventBus/Application/Queries/UserAgeQueryValidator.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.EventBus.IntegrationTests.Application.Queries;
+namespace Masa.Framework.IntegrationTests.EventBus.Application.Queries;
 
 public class UserAgeQueryValidator: AbstractValidator<UserAgeQuery>
 {
