+++ test/Masa.Framework.IntegrationTests.EventBus/Application/UserHandler.cs
@@ -1,12 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-using Masa.EventBus.IntegrationTests.Application.Command;
-using Masa.EventBus.IntegrationTests.Application.Events;
-using Masa.EventBus.IntegrationTests.Application.Queries;
-using Masa.EventBus.IntegrationTests.Domain.Aggregate;
-
-namespace Masa.EventBus.IntegrationTests.Application;
+namespace Masa.Framework.IntegrationTests.EventBus.Application;
 
 public class UserHandler
 {
