+++ test/Masa.Framework.IntegrationTests.EventBus/Infrastructure/CustomDbContext.cs
@@ -1,9 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-using Masa.EventBus.IntegrationTests.Domain.Aggregate;
-
-namespace Masa.EventBus.IntegrationTests.Infrastructure;
+namespace Masa.Framework.IntegrationTests.EventBus.Infrastructure;
 
 public class CustomDbContext : MasaDbContext
 {
