+++ test/Masa.Framework.IntegrationTests.EventBus/Infrastructure/Extensions/DefaultPublisher.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.EventBus.IntegrationTests.Infrastructure.Extensions;
+namespace Masa.Framework.IntegrationTests.EventBus.Infrastructure.Extensions;
 
 public class DefaultPublisher : IPublisher
 {
