+++ test/Masa.Framework.IntegrationTests.EventBus/Infrastructure/Middleware/RecordMiddleware.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.EventBus.IntegrationTests.Infrastructure.Middleware;
+namespace Masa.Framework.IntegrationTests.EventBus.Infrastructure.Middleware;
 
 public class RecordMiddleware<TEvent> : Middleware<TEvent>
     where TEvent : IEvent
