+++ test/Masa.Framework.IntegrationTests.EventBus/TestBase.cs
@@ -1,7 +1,11 @@
 // Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.EventBus.IntegrationTests;
+using Masa.Framework.IntegrationTests.EventBus.Infrastructure;
+using Masa.Framework.IntegrationTests.EventBus.Infrastructure.Extensions;
+using Masa.Framework.IntegrationTests.EventBus.Infrastructure.Middleware;
+
+namespace Masa.Framework.IntegrationTests.EventBus;
 
 [TestClass]
 public class TestBase
