+++ test/Masa.Framework.IntegrationTests.EventBus/_Imports.cs
@@ -16,13 +16,10 @@ global using Masa.Contrib.Ddd.Domain.Repository.EFCore;
 global using Masa.Contrib.Dispatcher.Events;
 global using Masa.Contrib.Dispatcher.IntegrationEvents;
 global using Masa.Contrib.Dispatcher.IntegrationEvents.EventLogs.EFCore;
-global using Masa.EventBus.IntegrationTests.Application.Command;
-global using Masa.EventBus.IntegrationTests.Application.Events;
-global using Masa.EventBus.IntegrationTests.Application.Queries;
-global using Masa.EventBus.IntegrationTests.Domain.Aggregate;
-global using Masa.EventBus.IntegrationTests.Infrastructure;
-global using Masa.EventBus.IntegrationTests.Infrastructure.Extensions;
-global using Masa.EventBus.IntegrationTests.Infrastructure.Middleware;
+global using Masa.Framework.IntegrationTests.EventBus.Application.Command;
+global using Masa.Framework.IntegrationTests.EventBus.Application.Events;
+global using Masa.Framework.IntegrationTests.EventBus.Application.Queries;
+global using Masa.Framework.IntegrationTests.EventBus.Domain.Aggregate;
 global using Microsoft.EntityFrameworkCore;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.DependencyInjection.Extensions;
